Core: fixed parsing IPv4 address.
Previously, the ngx_inet_addr function returned the parsed IPv4
address to the caller via the function return. And notably, it
returned INADDR_NONE, whose value is identical to INADDR_BROADCAST,
when failing to parse the input string as a valid IPv4 address,
which made it impossible for us to distinguish whether the input
string is invalid or or a valid IPv4 broadcast address.

In this patch, we changed the signature of ngx_inet_addr to return
a status code and the parsed IPv4 address is returned via an out
parameter as the ngx_inet6_addr function does. Besides this, the
broadcast address check which was previously performed accidentally
by comparing the return value of ngx_inet_addr with INADDR_NONE
should be performed explicitly from now on.